requirements.txt
openpyxl==3.0.10
xlwings==0.28.5
pandas==1.5.1
seaborn==0.12.1
matplotlib==3.6.2
python-dateutil==2.8.2